refactor(search): decompose search-index.ts into focused modules
Extract 4 new modules from search-index.ts (1986 → 992 lines):

- fts-query.ts (75 lines): FTS5 query sanitization + regex constants
- rrf.ts (46 lines): Reciprocal Rank Fusion scoring algorithm
- search-helpers.ts (177 lines): pure data transforms (row mappers,
  filters, link extraction, FTS metadata builder)
- search-queries.ts (813 lines): all 15 query methods + vector search,
  taking a SearchQueryContext for db/embedder access

The SearchIndex public API is unchanged — consumers (server.ts,
search-tools.ts, prompt-definitions.ts, file-watcher.ts) need zero
import path changes. Types are exported from search-index.ts and
imported via `import type` by sibling modules (no circular runtime
dependencies). A generic `bind` helper wires the query context.
