+++ src/vault-mcp/search/search-queries.ts
@@ -1,0 +1,807 @@
+// ── Query methods extracted from search-index.ts ──────────────
+
+import Database from "better-sqlite3"
+import { DateTime } from "luxon"
+import type { Logger } from "../../logger.js"
+import { describeError } from "../../utils/describe-error.js"
+import { assertPathHasExtension } from "../../utils/assert-path-has-extension.js"
+import { sanitizeFtsQuery } from "./fts-query.js"
+import { computeRrfScores } from "./rrf.js"
+import {
+  rowToMetadata,
+  noteRowToSearchResult,
+  notePassesFilters,
+  buildSnippetFromChunkText,
+} from "./search-helpers.js"
+import type {
+  VectorHit,
+  SearchResult,
+  HybridSearchResult,
+  NoteMetadata,
+  NoteRow,
+  TagCount,
+  PropertyKeyInfo,
+  PropertyValueCount,
+  VaultStats,
+  SearchFilters,
+  BacklinkEntry,
+  OutgoingLinkEntry,
+} from "./search-index.js"
+import type { Embedder } from "./embedder.js"
+
+// ── Context ────────────────────────────────────────────────────
+
+export type SearchQueryContext = {
+  readonly db: Database.Database
+  readonly getDailyNotesFolder: () => string | null
+  readonly vector: {
+    readonly embedder: Embedder | undefined
+    readonly knnSearchStmt: Database.Statement | null
+    readonly selectNoteMetadataStmt: Database.Statement
+  }
+}
+
+// ── Vector search (internal) ───────────────────────────────────
+
+const vectorSearch = async (
+  ctx: SearchQueryContext,
+  params: { query: string; limit: number },
+  logger: Logger,
+): Promise<VectorHit[]> => {
+  const { embedder, knnSearchStmt } = ctx.vector
+  if (!embedder || !knnSearchStmt) return []
+
+  try {
+    const queryEmbedding = await embedder.embedText(params.query)
+    const rows = knnSearchStmt.all(
+      Buffer.from(
+        queryEmbedding.buffer,
+        queryEmbedding.byteOffset,
+        queryEmbedding.byteLength,
+      ),
+      params.limit,
+    ) as Array<{ note_path: string; chunk_text: string; distance: number }>
+
+    // Deduplicate to best chunk per note — rows are ordered by distance
+    // ascending, so the first occurrence of each path is the closest match.
+    const bestChunkPerNote = new Map<string, VectorHit>()
+    for (const row of rows) {
+      if (!bestChunkPerNote.has(row.note_path)) {
+        bestChunkPerNote.set(row.note_path, {
+          path: row.note_path,
+          distance: row.distance,
+          chunkText: row.chunk_text,
+        })
+      }
+    }
+
+    logger.info("vector search", {
+      query: params.query,
+      knnHits: rows.length,
+      uniqueNotes: bestChunkPerNote.size,
+    })
+    return [...bestChunkPerNote.values()]
+  } catch (error) {
+    logger.warn("vector search failed, falling back to FTS-only", {
+      error: describeError(error),
+    })
+    return []
+  }
+}
+
+// ── Full-text search ───────────────────────────────────────────
+
+export const fullTextSearch = (
+  ctx: SearchQueryContext,
+  params: { query: string; filters?: SearchFilters },
+  logger: Logger,
+): SearchResult[] => {
+  // Build WHERE clause dynamically: each filter appends a condition + its bind params
+  const conditions: string[] = []
+  const queryParams: unknown[] = []
+
+  conditions.push("notes_fts MATCH ?")
+  queryParams.push(sanitizeFtsQuery(params.query))
+
+  if (params.filters?.folder) {
+    conditions.push("n.path LIKE ?")
+    queryParams.push(`${params.filters.folder}/%`)
+  }
+
+  if (params.filters?.tags) {
+    for (const tag of params.filters.tags) {
+      conditions.push(
+        "EXISTS (SELECT 1 FROM json_each(n.tags) WHERE value = ?)",
+      )
+      queryParams.push(tag)
+    }
+  }
+
+  if (params.filters?.related) {
+    for (const relatedNote of params.filters.related) {
+      conditions.push(
+        "EXISTS (SELECT 1 FROM json_each(n.related) WHERE value = ?)",
+      )
+      queryParams.push(relatedNote)
+    }
+  }
+
+  if (params.filters?.type) {
+    conditions.push("n.type = ?")
+    queryParams.push(params.filters.type)
+  }
+
+  if (params.filters?.properties) {
+    for (const [key, value] of Object.entries(params.filters.properties)) {
+      conditions.push(`json_extract(n.properties, '$.' || ?) = ?`)
+      queryParams.push(key, value)
+    }
+  }
+
+  const limit = params.filters?.limit ?? 20
+  const snippetTokens = params.filters?.snippet_tokens ?? 30
+  // Opt-in: the leading callout is omitted by default to keep this hot-path
+  // result lean; callers triaging which note to open can request it.
+  const includeLeadingCallout = params.filters?.include_leading_callout ?? false
+  queryParams.push(limit)
+
+  const sql = `
+    SELECT n.path, n.title,
+           snippet(notes_fts, 2, '', '', '...', ${Number(snippetTokens)}) as snippet,
+           rank * -1 as score, n.tags, n.folder, n.type, n.created, n.mtime,
+           n.bytes${includeLeadingCallout ? ", n.leading_callout" : ""}
+    FROM notes_fts
+    JOIN notes n ON n.path = notes_fts.path
+    WHERE ${conditions.join(" AND ")}
+    ORDER BY rank
+    LIMIT ?
+  `
+  // rank * -1: FTS5 rank is negative (lower = better), negated for human-friendly scoring
+
+  try {
+    const rows = ctx.db.prepare(sql).all(...queryParams) as Array<
+      Pick<
+        NoteRow,
+        | "path"
+        | "title"
+        | "tags"
+        | "folder"
+        | "type"
+        | "created"
+        | "mtime"
+        | "bytes"
+      > & {
+        snippet: string
+        score: number
+        leading_callout?: string | null
+      }
+    >
+
+    const results: SearchResult[] = rows.map((row) =>
+      noteRowToSearchResult({
+        row,
+        snippet: row.snippet,
+        score: Number(row.score.toPrecision(4)),
+        includeLeadingCallout,
+      }),
+    )
+    logger.info("full text search", {
+      query: params.query,
+      resultCount: results.length,
+    })
+    return results
+  } catch (error) {
+    logger.warn("full text search failed", {
+      query: params.query,
+      error: describeError(error),
+    })
+    return []
+  }
+}
+
+// ── Hybrid search ──────────────────────────────────────────────
+
+/** Hybrid search — combines FTS5 keyword search with sqlite-vec vector
+ *  similarity via RRF fusion. Falls back to FTS-only silently when no
+ *  embeddings are available. */
+export const hybridSearch = async (
+  ctx: SearchQueryContext,
+  params: { query: string; filters?: SearchFilters },
+  logger: Logger,
+): Promise<HybridSearchResult> => {
+  const userLimit = params.filters?.limit ?? 20
+  const snippetTokens = params.filters?.snippet_tokens ?? 30
+  const includeLeadingCallout = params.filters?.include_leading_callout ?? false
+  const candidateLimit = Math.min(userLimit * 3, 100)
+
+  // Run FTS with inflated limit to give RRF enough candidates
+  const ftsResults = fullTextSearch(
+    ctx,
+    {
+      query: params.query,
+      filters: { ...params.filters, limit: candidateLimit },
+    },
+    logger,
+  )
+
+  // Attempt vector search — returns [] on any failure
+  const vectorHits = await vectorSearch(
+    ctx,
+    { query: params.query, limit: candidateLimit },
+    logger,
+  )
+
+  // FTS-only fallback when no vectors are available
+  if (vectorHits.length === 0) {
+    const fallbackResults = ftsResults.slice(0, userLimit)
+    logger.info("hybrid search", {
+      query: params.query,
+      searchMode: "fts",
+      resultCount: fallbackResults.length,
+    })
+    return { results: fallbackResults, search_mode: "fts" }
+  }
+
+  // Compute RRF scores from both ranked lists
+  const rrfScores = computeRrfScores({
+    ftsRanked: ftsResults,
+    vectorRanked: vectorHits,
+  })
+
+  // Index FTS results and vector hits by path for O(1) lookup
+  const ftsResultsByPath = new Map(
+    ftsResults.map((result) => [result.path, result]),
+  )
+  const vectorHitsByPath = new Map(vectorHits.map((hit) => [hit.path, hit]))
+
+  // Build the merged result set, ordered by RRF score
+  const mergedResults: SearchResult[] = []
+  for (const { path, score } of rrfScores) {
+    const ftsResult = ftsResultsByPath.get(path)
+    if (ftsResult) {
+      // Path found via FTS — use its metadata and snippet, replace score
+      mergedResults.push({ ...ftsResult, score })
+      continue
+    }
+
+    // Vector-only result — look up metadata from the notes table
+    const noteRow = ctx.vector.selectNoteMetadataStmt.get(path) as
+      | NoteRow
+      | undefined
+    if (!noteRow) continue
+
+    // Apply filters that FTS would have applied via SQL
+    if (params.filters && !notePassesFilters(noteRow, params.filters)) continue
+
+    const vectorHit = vectorHitsByPath.get(path)
+    const snippet = vectorHit
+      ? buildSnippetFromChunkText(vectorHit.chunkText, snippetTokens)
+      : ""
+
+    mergedResults.push(
+      noteRowToSearchResult({
+        row: noteRow,
+        snippet,
+        score,
+        includeLeadingCallout,
+      }),
+    )
+  }
+
+  logger.info("hybrid search", {
+    query: params.query,
+    searchMode: "hybrid",
+    ftsResults: ftsResults.length,
+    vectorHits: vectorHits.length,
+    mergedResults: mergedResults.length,
+    returnedResults: Math.min(mergedResults.length, userLimit),
+  })
+  return {
+    results: mergedResults.slice(0, userLimit),
+    search_mode: "hybrid",
+  }
+}
+
+// ── Discovery queries ──────────────────────────────────────────
+
+/** Finds notes with a specific tag. Supports hierarchical prefix matching. */
+export const searchByTag = (
+  ctx: SearchQueryContext,
+  params: { tag: string; exactMatch?: boolean; limit?: number },
+  logger: Logger,
+): NoteMetadata[] => {
+  const limit = params.limit ?? 20
+
+  const condition = params.exactMatch
+    ? "EXISTS (SELECT 1 FROM json_each(n.tags) WHERE value = ?)"
+    : "EXISTS (SELECT 1 FROM json_each(n.tags) WHERE value = ? OR value LIKE ? || '/%')"
+
+  const queryParams: unknown[] = params.exactMatch
+    ? [params.tag, limit]
+    : [params.tag, params.tag, limit]
+
+  const sql = `
+    SELECT path, title, tags, related, folder, type, created, mtime, properties, leading_callout, bytes
+    FROM notes n
+    WHERE ${condition}
+    ORDER BY mtime DESC
+    LIMIT ?
+  `
+
+  const rows = ctx.db.prepare(sql).all(...queryParams) as NoteRow[]
+  const results = rows.map(rowToMetadata)
+  logger.info("search by tag", {
+    tag: params.tag,
+    resultCount: results.length,
+  })
+  return results
+}
+
+/** Lists notes in a folder, optionally including subfolders. */
+export const searchByFolder = (
+  ctx: SearchQueryContext,
+  params: { folder: string; recursive?: boolean; limit?: number },
+  logger: Logger,
+): NoteMetadata[] => {
+  const recursive = params.recursive ?? true
+  const limit = params.limit ?? 20
+
+  const condition = recursive
+    ? "path LIKE ? || '/%'"
+    : "path LIKE ? || '/%' AND path NOT LIKE ? || '/%/%'"
+
+  const queryParams: unknown[] = recursive
+    ? [params.folder, limit]
+    : [params.folder, params.folder, limit]
+
+  const sql = `
+    SELECT path, title, tags, related, folder, type, created, mtime, properties, leading_callout, bytes
+    FROM notes
+    WHERE ${condition}
+    ORDER BY mtime DESC
+    LIMIT ?
+  `
+
+  const rows = ctx.db.prepare(sql).all(...queryParams) as NoteRow[]
+  const results = rows.map(rowToMetadata)
+  logger.info("search by folder", {
+    folder: params.folder,
+    resultCount: results.length,
+  })
+  return results
+}
+
+/** Returns all tags in the vault with their note counts. */
+export const listAllTags = (
+  ctx: SearchQueryContext,
+  _params: Record<string, never>,
+  logger: Logger,
+): TagCount[] => {
+  const sql = `
+    SELECT value as tag, COUNT(DISTINCT notes.path) as count
+    FROM notes, json_each(notes.tags)
+    GROUP BY value
+    ORDER BY count DESC
+  `
+  const results = ctx.db.prepare(sql).all() as TagCount[]
+  logger.info("listed all tags", { count: results.length })
+  return results
+}
+
+/** Returns recently modified or created notes, sorted by chosen timestamp. */
+export const recentNotes = (
+  ctx: SearchQueryContext,
+  params: { sort_by?: "created" | "modified"; limit?: number },
+  logger: Logger,
+): NoteMetadata[] => {
+  const sortBy = params.sort_by ?? "modified"
+  const limit = params.limit ?? 20
+
+  // "created IS NULL" sorts NULLs last in a DESC ordering (SQLite evaluates 0/1)
+  const orderClause =
+    sortBy === "created"
+      ? "ORDER BY created IS NULL, created DESC"
+      : "ORDER BY mtime DESC" // SQL column is still `mtime`
+
+  const sql = `
+    SELECT path, title, tags, related, folder, type, created, mtime, properties, leading_callout, bytes
+    FROM notes
+    ${orderClause}
+    LIMIT ?
+  `
+
+  const rows = ctx.db.prepare(sql).all(limit) as NoteRow[]
+  const results = rows.map(rowToMetadata)
+  logger.info("recent notes", { sortBy, resultCount: results.length })
+  return results
+}
+
+// ── Property queries ───────────────────────────────────────────
+
+/** Returns all frontmatter property keys with note counts and top 3 sample values. */
+export const listPropertyKeys = (
+  ctx: SearchQueryContext,
+  params: { folder?: string },
+  logger: Logger,
+): PropertyKeyInfo[] => {
+  const folderCondition = params.folder
+    ? "WHERE n.path LIKE @folder || '/%'"
+    : ""
+
+  const keySql = `
+    SELECT property.key, COUNT(DISTINCT n.path) as count
+    FROM notes n, json_each(n.properties) property
+    ${folderCondition}
+    GROUP BY property.key
+    ORDER BY count DESC
+  `
+  const keySqlParams: Record<string, string> = {}
+  if (params.folder) keySqlParams.folder = params.folder
+  const keyRows = ctx.db.prepare(keySql).all(keySqlParams) as Array<{
+    key: string
+    count: number
+  }>
+
+  const sampleFolderCondition = params.folder
+    ? "AND path LIKE @folder || '/%'"
+    : ""
+
+  // For each key, fetch the 3 most common values as samples.
+  // json_array() wraps scalars so json_each works uniformly for
+  // both scalar ("active") and array (["a","b"]) property values.
+  const sampleSql = `
+    SELECT element.value, COUNT(*) as count
+    FROM (
+      SELECT properties FROM notes
+      WHERE json_type(properties, '$.' || @key) IS NOT NULL
+      ${sampleFolderCondition}
+    ) filtered, json_each(
+      CASE json_type(filtered.properties, '$.' || @key)
+        WHEN 'array' THEN json_extract(filtered.properties, '$.' || @key)
+        ELSE json_array(json_extract(filtered.properties, '$.' || @key))
+      END
+    ) element
+    WHERE typeof(element.value) IN ('text', 'integer', 'real')
+    GROUP BY element.value
+    ORDER BY count DESC
+    LIMIT 3
+  `
+  const sampleStmt = ctx.db.prepare(sampleSql)
+
+  const results: PropertyKeyInfo[] = keyRows.map((keyRow) => {
+    const sqlParams: Record<string, string> = { key: keyRow.key }
+    if (params.folder) sqlParams.folder = params.folder
+    const sampleRows = sampleStmt.all(sqlParams) as Array<{
+      value: string
+    }>
+    return {
+      key: keyRow.key,
+      count: keyRow.count,
+      sample_values: sampleRows.map((sampleRow) => String(sampleRow.value)),
+    }
+  })
+
+  logger.info("listed property keys", { count: results.length })
+  return results
+}
+
+/** Returns distinct values for a given property key with note counts. */
+export const listPropertyValues = (
+  ctx: SearchQueryContext,
+  params: { key: string; folder?: string; limit?: number },
+  logger: Logger,
+): PropertyValueCount[] => {
+  const limit = params.limit ?? 50
+  const folderCondition = params.folder ? "AND path LIKE @folder || '/%'" : ""
+
+  // json_array() wraps scalars so json_each works uniformly for
+  // both scalar ("active") and array (["a","b"]) property values.
+  const sql = `
+    SELECT element.value, COUNT(*) as count
+    FROM (
+      SELECT properties FROM notes
+      WHERE json_type(properties, '$.' || @key) IS NOT NULL
+      ${folderCondition}
+    ) filtered, json_each(
+      CASE json_type(filtered.properties, '$.' || @key)
+        WHEN 'array' THEN json_extract(filtered.properties, '$.' || @key)
+        ELSE json_array(json_extract(filtered.properties, '$.' || @key))
+      END
+    ) element
+    WHERE typeof(element.value) IN ('text', 'integer', 'real')
+    GROUP BY element.value
+    ORDER BY count DESC
+    LIMIT @limit
+  `
+
+  const sqlParams: Record<string, unknown> = { key: params.key, limit }
+  if (params.folder) sqlParams.folder = params.folder
+
+  const rows = ctx.db.prepare(sql).all(sqlParams) as Array<{
+    value: string | number
+    count: number
+  }>
+  const results = rows.map((row) => ({
+    value: String(row.value),
+    count: row.count,
+  }))
+  logger.info("listed property values", {
+    key: params.key,
+    count: results.length,
+  })
+  return results
+}
+
+/** Finds notes where a frontmatter property matches a value (exact match). */
+export const searchByProperty = (
+  ctx: SearchQueryContext,
+  params: { key: string; value: string; folder?: string; limit?: number },
+  logger: Logger,
+): NoteMetadata[] => {
+  const limit = params.limit ?? 20
+  const folderCondition = params.folder ? "AND n.path LIKE @folder || '/%'" : ""
+
+  // Two branches handle different property shapes:
+  // - Array properties (tags: ["a","b"]): check if @value is IN the array
+  // - Scalar properties (status: "active"): check direct equality
+  // Both branches CAST to TEXT for type-safe comparison (integer 4 = text "4")
+  const sql = `
+    SELECT path, title, tags, related, folder, type, created, mtime, properties, leading_callout, bytes
+    FROM notes n
+    WHERE (
+      (json_type(n.properties, '$.' || @key) = 'array'
+       AND EXISTS (
+         SELECT 1 FROM json_each(json_extract(n.properties, '$.' || @key))
+         WHERE CAST(value AS TEXT) = @value
+       ))
+      OR
+      (json_type(n.properties, '$.' || @key) IS NOT NULL
+       AND json_type(n.properties, '$.' || @key) != 'array'
+       AND CAST(json_extract(n.properties, '$.' || @key) AS TEXT) = @value)
+    )
+    ${folderCondition}
+    ORDER BY mtime DESC
+    LIMIT @limit
+  `
+
+  const sqlParams: Record<string, unknown> = {
+    key: params.key,
+    value: params.value,
+    limit,
+  }
+  if (params.folder) sqlParams.folder = params.folder
+
+  const rows = ctx.db.prepare(sql).all(sqlParams) as NoteRow[]
+  const results = rows.map(rowToMetadata)
+  logger.info("search by property", {
+    key: params.key,
+    value: params.value,
+    resultCount: results.length,
+  })
+  return results
+}
+
+// ── Link queries ───────────────────────────────────────────────
+
+/** Returns notes that link TO the given path (incoming links / backlinks). */
+export const getBacklinks = (
+  ctx: SearchQueryContext,
+  params: { path: string },
+  logger: Logger,
+): BacklinkEntry[] => {
+  assertPathHasExtension(params.path, ".md")
+  const sql = `
+    SELECT n.path, n.title, n.bytes
+    FROM links l
+    JOIN notes n ON n.path = l.source
+    WHERE l.target = ?
+    ORDER BY n.title
+  `
+  const rows = ctx.db.prepare(sql).all(params.path) as Array<{
+    path: string
+    title: string
+    bytes: number
+  }>
+  const results: BacklinkEntry[] = rows.map((row) => ({
+    path: row.path,
+    title: row.title,
+    bytes: row.bytes ?? 0,
+  }))
+  logger.info("get backlinks", {
+    path: params.path,
+    count: results.length,
+  })
+  return results
+}
+
+/** Returns notes and assets that the given path links TO (outgoing links).
+ *  Each entry carries a `kind` discriminator: "note" for .md targets,
+ *  "asset" for resolved non-markdown files (.canvas, .base, images, etc.),
+ *  defaulting to "note" for unresolved (broken) links. */
+export const getOutgoingLinks = (
+  ctx: SearchQueryContext,
+  params: { path: string },
+  logger: Logger,
+): OutgoingLinkEntry[] => {
+  assertPathHasExtension(params.path, ".md")
+  const sql = `
+    SELECT l.target as path,
+           n.title,
+           CASE WHEN n.path IS NOT NULL THEN 1
+                WHEN f.path IS NOT NULL THEN 1
+                ELSE 0 END as exists_flag,
+           CASE WHEN n.path IS NOT NULL THEN 'note'
+                WHEN f.path IS NOT NULL THEN 'asset'
+                ELSE 'note' END as kind,
+           n.bytes
+    FROM links l
+    LEFT JOIN notes n ON n.path = l.target
+    LEFT JOIN non_md_files f ON f.path = l.target
+    WHERE l.source = ?
+    ORDER BY l.target
+  `
+  const rows = ctx.db.prepare(sql).all(params.path) as Array<{
+    path: string
+    title: string | null
+    exists_flag: number
+    kind: "note" | "asset"
+    bytes: number | null
+  }>
+  const folder = ctx.getDailyNotesFolder()
+  const folderPrefix = folder !== null ? `${folder}/` : null
+  const results: OutgoingLinkEntry[] = rows.map((row) => ({
+    path: row.path,
+    title: row.title,
+    exists: row.exists_flag === 1,
+    kind: row.kind,
+    bytes: row.bytes ?? null,
+    daily_note_forward_ref:
+      row.exists_flag === 0 &&
+      folderPrefix !== null &&
+      row.path.startsWith(folderPrefix),
+  }))
+  logger.info("get outgoing links", {
+    path: params.path,
+    count: results.length,
+  })
+  return results
+}
+
+/** Finds notes with no incoming links (orphans). */
+export const findOrphans = (
+  ctx: SearchQueryContext,
+  params: { excludeFolders?: string[]; limit?: number },
+  logger: Logger,
+): NoteMetadata[] => {
+  const excludeFolders = params.excludeFolders ?? []
+  const limit = params.limit ?? 50
+
+  // One exclusion clause per folder, each bound to a positional parameter
+  const folderExclusions = Array(excludeFolders.length)
+    .fill("path NOT LIKE ? || '/%'")
+    .join(" AND ")
+  const whereClause = excludeFolders.length > 0 ? `AND ${folderExclusions}` : ""
+
+  // Self-links (source = target) are excluded from the backlink subquery
+  // so a note that only links to itself is still considered an orphan.
+  const sql = `
+    SELECT path, title, tags, related, folder, type, created, mtime, properties, leading_callout, bytes
+    FROM notes
+    WHERE path NOT IN (SELECT DISTINCT target FROM links WHERE source != target)
+      ${whereClause}
+    ORDER BY mtime DESC
+    LIMIT ?
+  `
+
+  const rows = ctx.db.prepare(sql).all(...excludeFolders, limit) as NoteRow[]
+  const results = rows.map(rowToMetadata)
+  logger.info("find orphans", { count: results.length })
+  return results
+}
+
+// ── Aggregate queries ──────────────────────────────────────────
+
+export type BrokenLinkResult = {
+  count: number
+  excludedFolder: string | null
+  excludedCount: number
+}
+
+/** Counts unique broken link targets — links whose targets exist in
+ *  neither the notes table nor the non_md_files table. When a daily
+ *  notes folder is configured, broken links under that folder are
+ *  excluded — they are forward-references (intentional "create on
+ *  click" navigation), not genuinely broken. Returns the count plus
+ *  exclusion metadata so callers can communicate what was filtered. */
+export const brokenLinkCount = (
+  ctx: SearchQueryContext,
+  _params: Record<string, never>,
+  logger: Logger,
+): BrokenLinkResult => {
+  const folder = ctx.getDailyNotesFolder()
+
+  if (folder === null) {
+    const row = ctx.db
+      .prepare(
+        `SELECT COUNT(DISTINCT target) as count
+         FROM links
+         WHERE target NOT IN (SELECT path FROM notes)
+           AND target NOT IN (SELECT path FROM non_md_files)`,
+      )
+      .get() as { count: number }
+    logger.info("broken link count", { count: row.count })
+    return { count: row.count, excludedFolder: null, excludedCount: 0 }
+  }
+
+  const folderPrefix = `${folder}/`
+  const brokenTargets = ctx.db
+    .prepare(
+      `SELECT DISTINCT target
+       FROM links
+       WHERE target NOT IN (SELECT path FROM notes)
+         AND target NOT IN (SELECT path FROM non_md_files)`,
+    )
+    .all() as Array<{ target: string }>
+
+  const count = brokenTargets.filter(
+    (row) => !row.target.startsWith(folderPrefix),
+  ).length
+  const excludedCount = brokenTargets.length - count
+
+  logger.info("broken link count", {
+    count,
+    dailyNotesFolder: folder,
+    excludedForwardRefs: excludedCount,
+  })
+  return { count, excludedFolder: folder, excludedCount }
+}
+
+/** Returns notes whose filesystem mtime falls within a calendar date
+ *  (server-local day boundaries, governed by the TZ env var). */
+export const modifiedOnDate = (
+  ctx: SearchQueryContext,
+  params: { date: string; limit?: number },
+  logger: Logger,
+): NoteMetadata[] => {
+  const limit = params.limit ?? 50
+  const dayStart = DateTime.fromISO(params.date)
+  const dayEnd = dayStart.plus({ days: 1 })
+
+  const sql = `
+    SELECT path, title, tags, related, folder, type, created, mtime, properties, leading_callout, bytes
+    FROM notes
+    WHERE mtime >= ? AND mtime < ?
+    ORDER BY mtime DESC
+    LIMIT ?
+  `
+  const rows = ctx.db
+    .prepare(sql)
+    .all(dayStart.toMillis(), dayEnd.toMillis(), limit) as NoteRow[]
+  const results = rows.map(rowToMetadata)
+  logger.info("modified on date", {
+    date: params.date,
+    resultCount: results.length,
+  })
+  return results
+}
+
+/** Lightweight aggregate counts — total notes, untagged notes, notes without
+ *  frontmatter properties. Single SQL to avoid multiple round-trips. */
+export const vaultStats = (
+  ctx: SearchQueryContext,
+  _params: Record<string, never>,
+  logger: Logger,
+): VaultStats => {
+  // Conditional aggregation: count all rows, then conditionally count rows
+  // whose tags/properties are the empty-JSON sentinel set by upsertNote.
+  const sql = `
+    SELECT
+      COUNT(*) as totalNotes,
+      COALESCE(SUM(CASE WHEN tags = '[]' THEN 1 ELSE 0 END), 0) as untaggedNotes,
+      COALESCE(SUM(CASE WHEN properties = '{}' THEN 1 ELSE 0 END), 0) as noPropertiesNotes
+    FROM notes
+  `
+  const row = ctx.db.prepare(sql).get() as VaultStats
+  logger.info("vault stats", row)
+  return row
+}
